SDC-3371: Refactore JDBC Source to not use ResultSet.isAfterLast() call
It seems that some DBs like Informatica and DB2 have trouble with the
ResultSet.isAfterLast call. They either automatically close the
ResultSet when it reaches end, which leads to an exception or do some
other questionable stuff.

This patch refactore our origin to not depend on this call and rather
keep state if we have reached end of the result set or not.

Change-Id: I4a2b14320073048ce50fdc5da9a9023e1f9d9c2b
Reviewed-on: https://review.streamsets.net/3388
Reviewed-by: Jonathan Natkins <natty@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>