SDC-3397. Field Renamer does not work as expected with arrays.
After the introduction of regex in fromField Names, the regex are
working propertly, but there was a regression which caused multiple
from Field expression having the constant array indexes (/list[0], /list[1]).
This is because the regex matched replaced all occurences of numbers
with \d+ which caused multiple fromExpressions to match the same fields
throwing exception, Users will use wildcard to specify all the fields
for eX: /list[*] if they want to rename the whole array/ migrate array to
individual fields.

This fix specifically handles the bug by not replacing constant
array indices with \d+ instead /list[0] will be patched as /list\[0\]
just escaping the array indices.

Change-Id: I8481df49716206f69aece830d7ef27a73dcb629d
Reviewed-on: https://review.streamsets.net/3449
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>