SDC-3449: Run CSD validation as part of gerrit run
I've moved the specific profile "cloudera" that is responsible for
building csd and parcel to cloudera-integration maven module and
included only the parcel creation there. The CSD will be now build in a
normal execution and hence it will be build in gerrit as well. Since CSD
build is extremely fast (just moving few files around) and running
validation routing, it shouldn't be an issue.

Change-Id: I8e15828a699cea1457a06f8151d4142773054b3e
Reviewed-on: https://review.streamsets.net/3563
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alejandro Abdelnur <tucu@streamsets.com>